requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
scikeras
pyyaml
h5py
xgboost
#-e .